{{cookiecutter.service_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.service_name}}"
version = "0.1.0"
description = "{{cookiecutter.description}}"
authors = ["{{cookiecutter.authors}}"]

[tool.poetry.dependencies]
python = ">={{cookiecutter.python_version}}"
fastapi = {extras = ["all"], version = "^0.73.0"}
pydantic = "^1.9.0"
pymongo = "^4.0.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-dotenv = "^0.5.2"
black = "^21.12b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
